requirements.txt
beautifulsoup4
descartes
duckdb
Fiona
flake8
geopandas 
matplotlib #==3.1.3
numpy #==1.18.5
pandas #==1.0.5
pyproj #==2.6.1.post1
requests
Rtree #==0.9.4
Shapely #==1.7.0
urllib3
zipp
feather-format
pyyaml
gptables
openpyxl
sphinx_rtd_theme
convertbng
pre-commit
google-cloud-storage
google-auth
mkdocs
